+++ src/oci/ai_speech/models/update_transcription_job_details.py
@@ -109,7 +109,7 @@ class UpdateTranscriptionJobDetails(object):
         """
         Gets the freeform_tags of this UpdateTranscriptionJobDetails.
         Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
-        Example: `{\"bar-key\": \"value\"}`
+        Example: `{\"bar-key\": \"value\"}`.
 
 
         :return: The freeform_tags of this UpdateTranscriptionJobDetails.
@@ -122,7 +122,7 @@ class UpdateTranscriptionJobDetails(object):
         """
         Sets the freeform_tags of this UpdateTranscriptionJobDetails.
         Simple key-value pair that is applied without any predefined name, type or scope. Exists for cross-compatibility only.
-        Example: `{\"bar-key\": \"value\"}`
+        Example: `{\"bar-key\": \"value\"}`.
 
 
         :param freeform_tags: The freeform_tags of this UpdateTranscriptionJobDetails.
@@ -135,7 +135,7 @@ class UpdateTranscriptionJobDetails(object):
         """
         Gets the defined_tags of this UpdateTranscriptionJobDetails.
         Defined tags for this resource. Each key is predefined and scoped to a namespace.
-        Example: `{\"foo-namespace\": {\"bar-key\": \"value\"}}`
+        Example: `{\"foo-namespace-1\": {\"bar-key-1\": \"value-1\", \"bar-key-2\": \"value-2\"}, \"foo-namespace-2\": {\"bar-key-1\": \"value-1\", \"bar-key-2\": \"value-2\"}}`.
 
 
         :return: The defined_tags of this UpdateTranscriptionJobDetails.
@@ -148,7 +148,7 @@ class UpdateTranscriptionJobDetails(object):
         """
         Sets the defined_tags of this UpdateTranscriptionJobDetails.
         Defined tags for this resource. Each key is predefined and scoped to a namespace.
-        Example: `{\"foo-namespace\": {\"bar-key\": \"value\"}}`
+        Example: `{\"foo-namespace-1\": {\"bar-key-1\": \"value-1\", \"bar-key-2\": \"value-2\"}, \"foo-namespace-2\": {\"bar-key-1\": \"value-1\", \"bar-key-2\": \"value-2\"}}`.
 
 
         :param defined_tags: The defined_tags of this UpdateTranscriptionJobDetails.